requirements.txt
streamlit
python-docx
PyMuPDF
pytesseract
opencv-python-headless
sentence-transformers
faiss-cpu
python-dotenv
cohere